Make debugging CSS a little easier
When debugging CSS from plugins and apps it is very hard to find a style with a problem, such as a typo. The inspectors in both Chrome and Firefox have length limits on what they can display in their windows, and having CSS blocks be one long string stops you from scrolling far enough to the right to see what the final output was. A newline between each block makes this inspection much easier. Some items are still too long, but at least they're separated into blocks so it is more likely you can find the style you are looking for.